requirements.txt
beautifulsoup4
certifi
console-menu
getpass
jsonmerge
jsonpatch
jsonpointer
jsonschema
numpy
oauthlib
pandas
pytz
PyYAML
requests
requests-oauthlib
selenium